[#256] split off shared superclass from CreateAppCmd and CreateProfileCmd; merge b986278c5b6089f1783b742a6e0ae2d41b7df270
If CreateProfileCommand extends CreateAppCommand, all options and positional parameters for `create-app` are *added* to subclass `create-profile`. The previous class hierarchy resulted in `create-profile` having two `NAME` positional parameters, which is incorrect.
